requirements.txt
PyYAML==6.0.2
pycryptodome
jinja2